refactor: extract _check_backend_version to _internal/_backend_version
Moves the function out of client.py into a dedicated internal module so
async_client.py can import it without depending on the client module.
